Audit #228/#229/#230 function sets; fix string-index and signedness bugs
Corrects three retail-evidenced defect classes found while reviewing the
91-function recovery set against retail:

- UseItem/BindCharacterItems: gppStringList element indices were divided
  by four again (e.g. gppStringList[0x164/4] read index 0x59, not 0x164).
  Rewritten in the file's byte-offset convention (0x590/4 etc.) so the
  correct localized notices are selected. BindCharacterItems now exact.
- SwapWeaponSetSlots0051D3B0: return type was unsigned int; retail
  returns in AL and callers test AL. Now unsigned char.
- ApplyCharmToMonsterTarget/ResolveCharmRefusal: spurious
  static_cast<unsigned int> casts turned the authored signed division
  (promoted unsigned char pair) into unsigned SHR; retail emits
  CDQ/SUB/SAR. Casts removed.

Remaining mismatches in the set reviewed instruction-by-instruction and
classified as compiler lowering: FCOMP scheduling, block ordering,
SEH-handler relocation, shared-epilogue external edges, and the known
UseItem/GetItemDisplayName retail inlining divergence.

+++ src/wiz8/local_code/Magic Effects.cpp
@@ -2474,9 +2474,7 @@ void ApplyCharmToMonsterTarget(W8SpellEffectEntry* effect)
     monster_info = MonsterInfoFromID(0xad1, MAGIC_EFFECTS_CPP, target->iMonsterID, 1);
     if (effect->Source.fBackfire != 0) {
         monster = GetMonsterDataForInfo(monster_info);
-        if (Random(100) < (static_cast<unsigned int>(monster->attribute_values_d1[4]) +
-                           static_cast<unsigned int>(monster->attribute_values_d1[1])) /
-                              2) {
+        if (Random(100) < (monster->attribute_values_d1[4] + monster->attribute_values_d1[1]) / 2) {
             QueueNpcScriptLine(0x16, 0, 0, 0);
         }
         return;
@@ -2541,9 +2539,7 @@ void ResolveCharmRefusal(W8SpellEffectEntry* effect)
         QueueDialogueNpcRefusal00576DA0();
     } else {
         monster = GetMonsterDataForInfo(monster_info);
-        if (Random(100) < (static_cast<unsigned int>(monster->attribute_values_d1[4]) +
-                           static_cast<unsigned int>(monster->attribute_values_d1[1])) /
-                              2) {
+        if (Random(100) < (monster->attribute_values_d1[4] + monster->attribute_values_d1[1]) / 2) {
             QueueNpcScriptLine(0x16, 0, 0, 0);
             return;
         }
